fix midnight bug

+++ Hang_man/Hang_man_V2/function.c
@@ -30,7 +30,7 @@ int Getlines()
     return(i);
 }
 
-char *giveWord(int a)
+char *giveWord(long a)
 {
     int file;
     int d;
